+++ sdk/computeschedule/arm-computeschedule/samples/v1-beta/typescript/src/scheduledActionsVirtualMachinesGetOperationStatusSample.ts
@@ -1,0 +1,52 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { ComputeScheduleClient } from "@azure/arm-computeschedule";
+import { DefaultAzureCredential } from "@azure/identity";
+
+/**
+ * This sample demonstrates how to virtualMachinesGetOperationStatus: Polling endpoint to read status of operations performed on virtual machines
+ *
+ * @summary virtualMachinesGetOperationStatus: Polling endpoint to read status of operations performed on virtual machines
+ * x-ms-original-file: 2025-04-15-preview/ScheduledActions_VirtualMachinesGetOperationStatus_MaximumSet_Gen.json
+ */
+async function scheduledActionsVirtualMachinesGetOperationStatusMaximumSet(): Promise<void> {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "CB26D7CB-3E27-465F-99C8-EAF7A4118245";
+  const client = new ComputeScheduleClient(credential, subscriptionId);
+  const result = await client.scheduledActions.virtualMachinesGetOperationStatus(
+    "ntfcikxsmthfkdhdcjpevmydzu",
+    {
+      operationIds: ["b211f086-4b91-4686-a453-2f5c012e4d80"],
+      correlationId: "b211f086-4b91-4686-a453-2f5c012e4d80",
+    },
+  );
+  console.log(result);
+}
+
+/**
+ * This sample demonstrates how to virtualMachinesGetOperationStatus: Polling endpoint to read status of operations performed on virtual machines
+ *
+ * @summary virtualMachinesGetOperationStatus: Polling endpoint to read status of operations performed on virtual machines
+ * x-ms-original-file: 2025-04-15-preview/ScheduledActions_VirtualMachinesGetOperationStatus_MinimumSet_Gen.json
+ */
+async function scheduledActionsVirtualMachinesGetOperationStatusMinimumSet(): Promise<void> {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "CB26D7CB-3E27-465F-99C8-EAF7A4118245";
+  const client = new ComputeScheduleClient(credential, subscriptionId);
+  const result = await client.scheduledActions.virtualMachinesGetOperationStatus(
+    "ykvvjfoopmkwznctgaiblzvea",
+    {
+      operationIds: ["duhqnwosjzexcfwfhryvy"],
+      correlationId: "b211f086-4b91-4686-a453-2f5c012e4d80",
+    },
+  );
+  console.log(result);
+}
+
+async function main(): Promise<void> {
+  await scheduledActionsVirtualMachinesGetOperationStatusMaximumSet();
+  await scheduledActionsVirtualMachinesGetOperationStatusMinimumSet();
+}
+
+main().catch(console.error);